fix: Only load view state after exiting tutorial
